requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
Django==5.2.1
django-debug-toolbar==5.2.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.0
iniconfig==1.1.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.10.3
pip==25.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg==3.2.9
pytest==8.3.4
python-decouple==3.8
setuptools==78.1.1
sqlparse==0.5.3
tomli==2.0.1
typing_extensions==4.12.2
wheel==0.45.1
